Updated fetch file and database
Updated the fetch_master_data.py with code to generate latitude and longitude coordinates from post codes when an entry didn't have them originally. Furthermore, add a column to the database that tags entries that used an approximate location from a post code. This allows plotting on the Folium map to include the tag in the popup.